requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-aliases==1.0.1
click-default-group==1.2.2
click-didyoumean==0.3.0
click-help-colors==0.9.1
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-appconf==1.0.5
django-celery-beat==2.2.1
django-celery-email==3.0.0
django-celery-results==2.2.0
django-environ==0.8.1
django-timezone-field==4.2.1
greenlet==1.1.2
gunicorn==20.1.0
honcho==1.1.0
kombu==5.2.2
Pillow==8.4.0
prompt-toolkit==3.0.23
psycopg2-binary==2.9.1
Pygments==2.10.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
redis==3.5.3
rich==9.13.0
six==1.16.0
SQLAlchemy==1.4.28
sqlparse==0.4.2
tabulate==0.8.9
typing-extensions==3.10.0.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
